requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
firebase-admin==6.5.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
gunicorn==21.2.0
httpx==0.27.0
aiohttp==3.9.5
requests==2.31.0